+++ apps/frontend/app/lib/hooks.ts
@@ -3,7 +3,7 @@ import { useMantineTheme } from "@mantine/core";
 import { useNavigate, useSearchParams } from "@remix-run/react";
 import { useAtom } from "jotai";
 import Cookies from "js-cookie";
-import { COOKIES_KEYS, getStringAsciiValue } from "./generals";
+import { ApplicationKey, getStringAsciiValue } from "./generals";
 import { InProgressWorkout, currentWorkoutAtom } from "./workout";
 
 export function useGetMantineColor() {
@@ -44,7 +44,7 @@ export function getWorkoutStarter() {
 
 	const fn = (wkt: InProgressWorkout) => {
 		setCurrentWorkout(wkt);
-		Cookies.set(COOKIES_KEYS.isWorkoutInProgress, "true", {
+		Cookies.set(ApplicationKey.CurrentWorkout, "true", {
 			expires: 2,
 			sameSite: "Strict",
 			secure: true,