+++ apps/frontend/app/routes/_dashboard.tsx
@@ -45,7 +45,7 @@ import { match } from "ts-pattern";
 import { joinURL } from "ufo";
 import { redirectIfNotAuthenticated } from "~/lib/api.server";
 import { colorSchemeCookie } from "~/lib/cookies.server";
-import { LOCAL_STORAGE_KEYS, getLot } from "~/lib/generals";
+import { ApplicationKey, getLot } from "~/lib/generals";
 import { getCoreDetails, getUserPreferences } from "~/lib/graphql.server";
 import classes from "~/styles/dashboard.module.css";
 
@@ -149,7 +149,7 @@ export default function Layout() {
 		  }
 		| undefined
 	>({
-		key: LOCAL_STORAGE_KEYS.savedOpenedLinkGroups,
+		key: ApplicationKey.SavedOpenedLinkGroups,
 		defaultValue: {
 			fitness: false,
 			media: false,